requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.5.2
arrow
backports.zoneinfo==0.2.1
boto3==1.6.19
botocore==1.9.23
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
docutils==0.20.1
exceptiongroup==1.3.0
fastapi==0.121.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
httpx==0.27.0
idna==3.11
jmespath==0.10.0
loguru==0.7.3
mysql-connector-python
numpy==1.24.4
protobuf==5.29.5
PyAudio==0.2.14
pycparser==2.23
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pymysql
python-dateutil
python-dotenv==1.0.1
python-multipart
redis
requests==2.32.4
s3transfer==0.1.13
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
starlette==0.44.0
typing-extensions==4.13.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.33.0
websockets==13.1
win32-setctime==1.2.0